Create app.py
Initially I have imported json as I am going to use json as a database for the cars.

I have also created a cars json array of the information provided in the brief.

Finally I created a quick for loop to test the json file if it works.

Everything works well so far.