basic structure
We have created a function called carById that takes in an id.

When the user enters option 1, we ask the user for input, the id of the car to get.
We have a try block to validate if the input is a number.
- If the input is a number we pass the id to the carById function and print the id (test)
- if the input is not a number then we display a message.

Right now the message says that the id entered does not exist or is not a number. We might want to check if the id exists with a loop. This will most likely be in the carById function.