updated carById
In the carById function we now have a for loop that iterates through the cars in the json.
Within the for loop, we have an if statement that takes the current cars id and checks if it is the same as the id entered by the user. If the id is correct we will print out all the information except chassis_no and then break out of the for loop.

At the end of the loop we will display a message if a car with the passed in id was not found.