new submenu
We now have created a submenu that will display the makes of cars in the cars json.

The user will be able to enter the index of the make of the car and that will be then used to iterate through the cars json to get prices of the make.

Now we need to implement the maths portion of the function